libcap: avoid %optflags macro
%optflags is the result of augmenting %__global_compiler_flags with
arch-specific flags. These may not work when cross-compiling as the
target arch may not match the host arch.

Switch to the baseline %__global_compiler_flags instead. These flags
are only relevant when compiling software for the host, which should
not be included in the final set of packaged artifacts, so any missed
optimizations are not a concern.

Signed-off-by: Ben Cressey <bcressey@amazon.com>