systemd: refactor to list packaged files
The current wildcard-based file includes make it harder to review the
impact of changing the build time configuration to enable features.

For optional functionality, it also makes it more tedious to place
files into the correct subpackage, since each file has to be excluded
from the main package and then also listed in the subpackage. It's
easy to miss a related file and then it ends up packaged twice or not
moved to the subpackage.

Remove all wildcard includes and replace with lists of files. While
we're at it, remove a handful of units that are unlikely to ever be
useful for Bottlerocket.

Signed-off-by: Ben Cressey <bcressey@amazon.com>