systemd: enable cryptsetup and tpm functionality
Enable the systemd features related to cryptsetup and TPM 2.0 usage.

For full functionality, this also requires making the build aware
that OpenSSL is available, which in turn enables more features in the
`systemd-repart`, `systemd-dissect`, and `systemd-resolved` programs
that were already being built.

`systemd-dissect` needs a function that AWS-LC doesn't yet support in
the FIPS module, and `systemd-resolved` enables DNSSEC in spite of
the other build options saying that DNSSEC should be disabled. The
two new patches suppress these uses of OpenSSL.

In the absence of DNSSEC, there are no users of TLS, so the third
patch adjusts the dependency to just libcrypto and not also libssl.

`systemd-repart` and `systemd-creds` pick up some new features. The
libsystemd shared library is now linked with libcrypto, and dlopens
libcryptsetup and libtss2 as needed if those features are used.

Add runtime requirements on these new libraries to the base systemd
package, to ensure that all the new functionality works as expected.

For now, move the programs and targets related to cryptsetup into a
new subpackage, since they aren't needed for all variants.

Co-authored-by: Mikael Knutsson (mikn) <mikael.knutsson@gmail.com>
Co-authored-by: Ben Cressey <bcressey@amazon.com>