igzip: optimize pipe sizes for containerd
Unlike pigz, igzip does not have dedicated threads for reading and
writing to copy to or from the working buffer during decompression.
If either the input or the output is a pipe, this can lead to delays
in decompression if the pipe is left to its 64 KiB default size.

Attempt to increase the pipe size to match the 1 MiB buffer size.
This is the default value of `fs.pipe-max-size`, so it's reasonably
likely to work on most systems. Ignore any errors since the program
works with the smaller buffer size, just not as efficiently.

Signed-off-by: Ben Cressey <bcressey@amazon.com>