sources: add brush
Various third-party programs might expect to execute programs via
system(3) or similar library functions. In many cases, this can be
worked around with a small patch, like the one to `open-vm-tools`, to
switch to a different library function. This works best for executing
a bounded set of known programs, since that avoids any possibility of
arbitrary code execution.

Both versions of the AWS SDK for Go will invoke `sh -c '...'` for an
AWS config file that contains a `credential_process` directive. Since
we don't control the contents of that config file, we can't enforce
that only specific credential processes are included. Moreover, quite
a few of our Go-based packages vendor the Go SDK, and patching all of
them would create a maintenance chore.

brush solves this problem by providing just enough of a `sh` wrapper
to handle argument parsing and to check whether the requested program
is allowed to run. It can be extended through the conventional means
of adding symlinks to a specific directory, which permits downstream
packages to allow other programs without affecting the global set of
what's allowed to run for all variants.

brush also supports an optional configuration file for each program,
to ensure that required arguments are present and blocked arguments
are not.

Signed-off-by: Ben Cressey <bcressey@amazon.com>