aws-signing-helper: depend on a program for sh
aws-signing-helper can be specified as the credential process for a
profile in the AWS config file. Consequently, it needs a program to
provide an `sh -c '...'` interface.

Add symlinks to the allowed programs directory for brush, so that if
brush is selected as the provider - which it will be in most cases -
then the helper will be permitted to run.

Also add a configuration file, so that only the "credential-process"
subcommand can be invoked this way, and to block "--debug" since that
can cause signed requests to be logged.

Signed-off-by: Ben Cressey <bcressey@amazon.com>