nvidia-container-toolkit: fix ldcache parsing for aarch64
The architecture flag for aarch64 is currently missing from the
supported architecture flags list. This omission causes the getEntries
function to exclude all libraries found on aarch64 hosts. As a result,
helper programs like nvidia-ctk are unable to generate CDI
specifications for the aarch64 architecture.

This fix adds the missing aarch64 architecture flag, using the same
value as defined in libnvidia-container[1], which maintains a more
comprehensive list of supported architectures.

[1]: https://github.com/NVIDIA/libnvidia-container/blob/a198166e1c1166f4847598438115ea97dacc7a92/src/ldcache.h#L21

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>