ghostdog: Add Infiniband GUID writing support
The latest Fabric Manager uses a new tool called nvlsm to manage
nvswitches. This requires communication over an Infiniband device and
both Fabric Manager and NVLSM need to be configured to use the same Port
GUID. This adds a new command line argument to ghostdog that detects if
the infiniband devices are present, and if so, finds a suitable GUID for
communication and writes this to the requested environment file. If there
are no devices present, it exits without writing the file.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>