nvidia-container-runtime: skip detecting mode for prestart hook
The prestart hook attempts to detect the mode used by the NVIDIA
container runtime, and fails if the detected mode is different than
'legacy'. This breaks the NVIDIA legacy runtime when the default mode is
set to 'cdi'.

Prevent the prestart hook from detecting the runtime mode, as using
'cdi' as the default mode while using the NVIDIA legacy runtime is a
valid configuration.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>