release: switch targets with "start" vs. "isolate"
"isolate" implies that the units needed by the new target should be
started, and that any units that aren't needed should be stopped. The
latter can be the source of surprises in the presence of units that
start dynamically in response to certain events.

"start" achieves the intent of activating a new set of units, which
is sufficient for Bottlerocket's needs, and should be more efficient
and less error-prone.

Signed-off-by: Ben Cressey <bcressey@amazon.com>