containerd: configure soci-snapshotter
Configure containerd to optionally specify soci-snapshotter. This is
done via imports of templated config files for containerd. The purpose
of these being defined in a separate file is to avoid modification of
the snapshotter at runtime; the template will be rendered only on boot.

Include patches for containerd-1.7 to bring its configuration import
logic to parity with containerd 2.x. Prior to the patches, entire
sections of plugin configuration would be overwritten by merges to other
settings in the same plugin configuration section.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>