release: add subpackage to enable swap
In low memory, pre-OOM-kill conditions, an orchestrator agent like
kubelet that is responsible for freeing up system resources may be
unable to terminate a process that allocates too much memory, if its
clean pages are dropped from the cache and must be read back from the
filesystem's block device whenever it runs, only to have those pages
dropped again when the scheduler selects another process to run.

For storage devices with metered I/O, such Amazon EBS gp2 volumes,
this cycle of reading the same filesystem blocks over and over can
exhaust the available credits, and leave the system livelocked for
for hours or days.

Providing a swap device gives the kernel an alternative to evicting
clean pages to free up memory prior to invoking the OOM killer: it
can swap out anonymous memory instead.

If the swap device were also stored on a device with I/O limits, the
same failure mode might arise when faulting in pages for anonymous
memory.

Using zram for the swap device avoids this problem. Instead of being
swapped to disk, the anonymous memory pages are compressed, and the
savings relative to the uncompressed size are returned to the system.
The zram device is not used until physical memory is exhausted; it
has an overhead of around 1%, or 1 MiB for the 1 GiB reservation.

Enabling zram-backed swap may become the default for Bottlerocket in
the future. For now, add the logic to a new subpackage so variants
can opt in by specifying "release-swap" instead of "release" in their
package list.

Signed-off-by: Ben Cressey <bcressey@amazon.com>