apiserver: add lockdown API
This API places the OS into "lockdown" mode, preventing any changes
to persisted settings until the next reboot.

By making it an API command, it can be triggered when settings are in
a known-good state: by bootstrap command; in a bootstrap container;
or by some other container later in boot.

Signed-off-by: Ben Cressey <bcressey@amazon.com>