release: measure boot phases into TPM2 device
The systemd-pcrphase-*.service units are derived from and inspired
by the upstream systemd units with the same names. Those are only
enabled if `systemd-stub` and unified kernel images (UKIs) are in
use, but the underlying idea is very clever and worth adopting.

Because `rottweiler` seals - by way of `systemd-creds` - the keys it
generates to the PCR where these measurements are sent, the keys can
only be unsealed in the future if the system has not advanced beyond
the phase of boot during which they were first generated.

For example, a key that's generated and sealed on first boot before
the system reaches "sysinit" - as the keys for local storage and the
datastore are - can never be unsealed after the "sysinit" phase on
the first boot or any other boot.

The local storage key is only needed for the initial attachment of
the device, and the datastore key only needs to be loaded into the
kernel once. Consequently, there's no loss in functionality from
making the keys inaccessible, only a gain in security.

Ephemeral storage can be initialized in several phases:
1. During preconfigured, by a bootstrap command
2. During configured, by a bootstrap container
3. During multi-user, by any other container

Since the ephemeral storage key is also sealed to the phase PCR,
it means that it can only ever be initialized during that phase. If
the initialization step is moved from "bootstrap container" to
"bootstrap command", then the PCR won't (yet) have reached the state
required to unlock it and the key couldn't be unsealed.

For compatibility with a potential migration to `systemd-stub` or
UKIs in the future, perform the same phase measurements that upstream
does, along with custom measurements for Bottlerocket's configuration
phases.

Signed-off-by: Ben Cressey <bcressey@amazon.com>