libaudit: add audit-rules subpackage
Create a new audit-rules subpackage containing the audit-rules.service
unit file and default audit.rules. This allows the audit rules to be
excluded from builds by overriding with a higher epoch package.

The subpackage is required by the systemd packages that provide auditd
so the rules are needed by that package.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>